Refactored notifications functionality
