+++ ansh-common/src/main/java/com/ansh/entity/animal/UserProfile.java
@@ -18,7 +18,7 @@ import java.util.Set;
 import lombok.Data;
 
 @Entity
-@Table(name = "userprofiles", schema = "public")
+@Table(name = "user_profiles", schema = "public")
 @Data
 public class UserProfile {
